requirements.txt
apache-tvm==0.11.1
apex==0.9.10.dev0
gin-config==0.5.0
hypothesis==6.102.6
iopath==0.1.10
ipython==8.25.0
matplotlib==3.9.0
onnx==1.16.1
pandas==2.2.2
pyre-extensions==0.0.30
scikit-learn==1.5.0
tensorboard==2.16.2
torchrec==0.7.0
urwid-readline==0.14
